refactor: move toolbar components to seperate file
